Common test implementation for resiliency config
