+++ api/tests/Cargo.toml
@@ -17,10 +17,9 @@ cmake.workspace = true
 azihsm_api.workspace = true
 azihsm_api_tests_macro.workspace = true
 azihsm_crypto.workspace = true
-fs2.workspace = true
+azihsm_resiliency_test_helpers.workspace = true
 hex.workspace = true
 libtest-mimic.workspace = true
-parking_lot.workspace = true
 tracing.workspace = true
 tracing-subscriber.workspace = true
 tracing-tree.workspace = true
