+++ tools/resiliency_stress/Cargo.toml
@@ -15,9 +15,9 @@ path = "src/main.rs"
 azihsm_api.workspace = true
 azihsm_crypto.workspace = true
 azihsm_ddi_sim.workspace = true
+azihsm_resiliency_test_helpers.workspace = true
 azihsm_res_test_dev = { optional = true, workspace = true }
 clap = { features = ["derive"], workspace = true }
-fs2.workspace = true
 libc.workspace = true
 memmap2 = "0.9"
 parking_lot = { features = ["deadlock_detection"], workspace = true }
